+++ src/main/java/baseball/util/InputGetter.java
@@ -1,4 +1,4 @@
-package baseball.utils;
+package baseball.util;
 
 import camp.nextstep.edu.missionutils.Console;
